correct file architecture and add vterminal
