+++ CMakeLists.txt
@@ -58,14 +58,14 @@ set_target_properties(tests
 )
 
 add_executable(2at2
-    src/stringPlus.c
-    src/parser.c
-    src/builder.c
-    src/debug.c
-    src/error.c
-    src/binExporter.c
-    src/2at2.c
-    src/ast.c
+  src/stringPlus.c
+  src/parser.c
+  src/builder.c
+  src/debug.c
+  src/error.c
+  src/binExporter.c
+  src/2at2.c
+  src/ast.c
 )
 
 target_include_directories(2at2
@@ -86,13 +86,12 @@ set_target_properties(2at2
 
 add_executable(vat2
     src/stringPlus.c
-    src/parser.c
-    src/builder.c
     src/debug.c
     src/error.c
-    src/binExporter.c
     src/vat2.c
-    src/ast.c
+    src/vProc.c
+    src/vMem.c
+    src/vTerminal.c
 )
 
 target_include_directories(vat2
