+++ src/vMem.c
@@ -1,0 +1,3 @@
+#include "vMem.h"
+
+char ram[16384][8];
