+++ src/vMem.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+#ifndef V_MEM_H
+#define V_MEM_H
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
